chore(home): Add Italian translations for home page
Several changes were made in the code to enable Italian translations for the home page. Text content in `it.json` were expanded with Italian translations for page titles, subtitles and other elements. Also, code in `home.component.ts` and `home.component.html` was updated to use the added translations instead of hardcoded text.