feat(restaurant picker): Add restaurant picker component and styles
Implemented a new restaurant picker component to enable users to choose between a restaurant and a pub. New internationalization keys for Italian language have been added for these options. Appropriate styles for the component have been set in LESS files. The selected option is then displayed on the restaurants page.

Signed-off-by: riccardocampani96 <riccardocampani96@gmail.com>