feat(restaurants): Implement loading and display of restaurant menus
This update removes the restaurants service and creates a new menus service that loads menus from a local JSON file. Menu data and restaurants are displayed on the restaurants component based on user selection. This also introduces model classes for dish and menus, changing "restaurant" and "pub" to "menù ristorante" and "menù birreria" respectively in italian translations.

Signed-off-by: riccardocampani96 <riccardocampani96@gmail.com>