style(restaurants): Add translations and improve UI for restaurants
This commit includes translations of restaurant categories in the i18n italian json. The UI for both the restaurant and events pages were enhanced to provide more detailed and organized information, and to improve responsiveness on multiple screen sizes. Additionally, the data representing the drinks in restaurants was refactored to be more intuitive and maintainable.

Signed-off-by: riccardocampani96 <riccardocampani96@gmail.com>