fix(nav item click): Rename MenuService to NavService and update usage
The code changes rename `MenuService` to `NavService` to reflect its purpose more accurately. This also required updating all the instances where `MenuService` was used to the new `NavService`. It also adds a new `navItemClick` method in the `NavComponent` class for handling navigation items click events and updates the sidebar menu rendering accordingly in the component template.

Signed-off-by: riccardocampani96 <riccardocampani96@gmail.com>